services/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
redis[hiredis]==5.1.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0
httpx==0.27.2
aiosmtplib==3.0.1
sendgrid==6.11.0
python-multipart==0.0.12
sentry-sdk[fastapi]==2.19.2
opentelemetry-sdk>=1.40.0
opentelemetry-exporter-otlp>=1.40.0
opentelemetry-instrumentation-fastapi>=0.49b0
opentelemetry-instrumentation-sqlalchemy>=0.49b0
opentelemetry-instrumentation-redis>=0.49b0
psycopg2-binary==2.9.9
stripe==8.10.0

# Test dependencies (also installed in CI)
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
aiosqlite==0.20.0
